Create Everything related to ArchiveInvalidation.h
This file contains full translations, from x86 to almost-C++, for every single subroutine in the game that is related to ArchiveInvalidation -- everything from the code to load each individual BSA based on sArchiveList and Plugins.txt, to the code that actually performs ArchiveInvalidation.

Think of this file as a "pidgin language" version of C++. Among other things, it's not strict about types and variable name reuse; my preference was for the code to be readable and recognizable to C++ developers, while still accurately indicating what registers are used for what in any case where I could do so without compromising readability.